folds dedicated _metrics files into _controller
